changed plotting function and removed typo
